Updated packages
Moved anything to do with auto complete (including
auto-complete-clang-async) outside the elpa so that they can be auto
loaded properly (they weren't before). Now everything is seamless as
long as /usr/bin/clang-complete exists. Not sure how its going to work
if that binary is not in place because I haven't tested it yet.